refact : Suggest 코드 분리
- 기존 suggest url을 restful 하게 수정
